Move gui-related metadata to a new file
